+++ services/web/server/tests/unit/with_dbs/06/test_projects_open_close.py
@@ -12,7 +12,6 @@ from unittest.mock import call
 
 import pytest
 import socketio
-from _helpers import ExpectedResponse, standard_role_response
 from aiohttp import web
 from aiohttp.test_utils import TestClient, TestServer
 from models_library.projects_access import Owner
@@ -26,6 +25,10 @@ from models_library.projects_state import (
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import log_client_in
 from pytest_simcore.helpers.utils_projects import assert_get_same_project
+from pytest_simcore.helpers.utils_webserver_projects import (
+    ExpectedResponse,
+    standard_role_response,
+)
 from servicelib.aiohttp.web_exceptions_extension import HTTPLocked
 from simcore_service_webserver.db_models import UserRole
 from simcore_service_webserver.projects.projects_handlers import (
