+++ services/web/server/tests/unit/with_dbs/conftest.py
@@ -25,12 +25,12 @@ import simcore_postgres_database.cli as pg_cli
 import simcore_service_webserver.db_models as orm
 import simcore_service_webserver.utils
 import sqlalchemy as sa
-from _helpers import MockedStorageSubsystem
 from _pytest.monkeypatch import MonkeyPatch
 from aiohttp import web
 from aiohttp.test_utils import TestClient, TestServer
 from pytest_simcore.helpers.utils_dict import ConfigDict
 from pytest_simcore.helpers.utils_login import NewUser
+from pytest_simcore.helpers.utils_webserver_projects import MockedStorageSubsystem
 from servicelib.aiohttp.application_keys import APP_DB_ENGINE_KEY
 from servicelib.common_aiopg_utils import DSN
 from simcore_service_webserver._constants import INDEX_RESOURCE_NAME
